requirements.txt
fastapi
uvicorn
streamlit
requests
python-docx
PyMuPDF
pydantic
langchain
langchain-community
langchain-openai
python-dotenv
python-multipart
pinecone
openai>=1.12.0
audio-recorder-streamlit